FastAPI/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
# -f https://download.pytorch.org/whl/torch_stable.html
fastapi[all]==0.63.0
uvicorn[standard]==0.13.4
gunicorn==20.0.4
-r https://raw.githubusercontent.com/snowflakedb/snowflake-connector-python/v2.6.2/tested_requirements/requirements_38.reqs
snowflake-connector-python
transformers
tensorflow-hub
tensorflow-cpu
# torch==1.5.0+cpu
# torchvision==0.6.0+cpu
# torchaudio